# minikaenguru-product/backend/mk-gateway/src/main/resources/META-INF/openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "mk-gateway-api",
    "description" : "API f�r die Minik�nguru-Anwendungen",
    "contact" : {
      "name" : "Heike Winkelvo�",
      "url" : "https://mathe-jung-alt.de",
      "email" : "info@egladil.de"
    },
    "license" : {
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.3.1"
  },
  "servers" : [ {
    "url" : "http://localhost:9510",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:9510",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/mk-gateway/admin/wettbewerbe/wettbewerb/{jahr}" : {
      "get" : {
        "tags" : [ "Admin Wettbewerb Resource" ],
        "parameters" : [ {
          "name" : "jahr",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/mk-gateway/mkbiza/wettbewerbe" : {
      "get" : {
        "tags" : [ "Mk Bi Za Resource" ],
        "summary" : "Gibt Infos über die API zurück.",
        "operationId" : "getInfos",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MkBiZaWettbewerb"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Serverfehler",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          }
        }
      }
    },
    "/mk-gateway/mkbiza/wettbewerbe/{jahr}" : {
      "get" : {
        "tags" : [ "Mk Bi Za Resource" ],
        "summary" : "Gibt die Statistik eines Wettbewerbs zurück",
        "operationId" : "getStatistikWettbewerb",
        "parameters" : [ {
          "name" : "jahr",
          "in" : "path",
          "description" : "Jahr des Wettbewerbs",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MkBiZaWettbewerbDetails"
                }
              }
            }
          },
          "401" : {
            "description" : "S2S-Autentifizierung schlug fehl",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          },
          "404" : {
            "description" : "Jahr existsiert nicht oder Wettbewerb ist noch nicht beendet",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          },
          "500" : {
            "description" : "Serverfehler",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          }
        }
      }
    },
    "/mk-gateway/mkbiza/wettbewerbe/{jahr}/{klassenstufe}/aufgaben/{nummer}" : {
      "get" : {
        "tags" : [ "Mk Bi Za Resource" ],
        "summary" : "Gibt die Statistik für eine spezielle Aufgabe zurück",
        "operationId" : "getStatistikAufgabe",
        "parameters" : [ {
          "name" : "jahr",
          "in" : "path",
          "description" : "Jahr des Wettbewerbs",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "klassenstufe",
          "in" : "path",
          "description" : "Klassenstufe",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Klassenstufe"
          }
        }, {
          "name" : "nummer",
          "in" : "path",
          "description" : "Nummer der Aufgabe",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MkBiZaWettbewerbDetails"
                }
              }
            }
          },
          "401" : {
            "description" : "S2S-Autentifizierung schlug fehl",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          },
          "404" : {
            "description" : "Jahr existsiert nicht oder Wettbewerb ist noch nicht beendet oder Aufgabennummer existsiert nicht",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          },
          "500" : {
            "description" : "Serverfehler",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessagePayload"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Klassenstufe" : {
        "enum" : [ "IKID", "EINS", "ZWEI" ],
        "type" : "string"
      },
...
      "MessagePayload" : {
        "type" : "object",
        "properties" : {
          "level" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "gesamtpunktverteilung"
        }
      },
      "MkBiZaGruppierungsitem" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "anzahl" : {
            "format" : "int64",
            "type" : "integer"
          }
        }
      },
      "MkBiZaWettbewerb" : {
        "type" : "object",
        "properties" : {
          "jahr" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "$ref" : "#/components/schemas/WettbewerbStatus"
          }
        }
      },
      "MkBiZaWettbewerbDetails" : {
        "type" : "object",
        "properties" : {
          "jahr" : {
            "format" : "int32",
            "type" : "integer"
          },
          "anzahlKinderGesamt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "anzahlPrivatanmeldungen" : {
            "format" : "int64",
            "type" : "integer"
          },
          "anzahlSchulanmeldungen" : {
            "format" : "int64",
            "type" : "integer"
          },
          "teilnehmendeSchulenGesamt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "klassenstufen" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Klassenstufe"
            }
          },
          "schulenJeLand" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MkBiZaGruppierungsitem"
            }
          },
          "schulkinderJeLand" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MkBiZaGruppierungsitem"
            }
          },
          "kinderJeTeilnahmeart" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MkBiZaGruppierungsitem"
            }
          },
          "kinderJeKlassenstufe" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MkBiZaGruppierungsitem"
            }
          },
          "kinderJeSprache" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MkBiZaGruppierungsitem"
            }
          }
        }
      },
...
      "WettbewerbStatus" : {
        "enum" : [ "ERFASST", "ANMELDUNG", "DOWNLOAD_LEHRER", "DOWNLOAD_PRIVAT", "BEENDET" ],
        "type" : "string"
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}